backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
pymupdf
python-docx
spacy
nltk
sentence-transformers
numpy
pandas
